cleaner architecture

+++ core/services/student_enrollment_service.py
@@ -8,17 +8,11 @@ class StudentEnrollmentService:
     def create_enrollment(user, data):
                 
         school = user.counselor.school        
-        student = Student.objects.for_user(user).get(id=data["student"])
-        school_year = SchoolYear.objects.get(id=data["school_year"])
-        
-        class_level = None
-        if data.get("class_level"):
-            class_level = ClassLevel.objects.get(id=data["class_level"])
 
         return StudentEnrollmentRepository.create(
-            student=student,
-            school_year=school_year,
-            class_level=class_level,
+            student=data["student"],
+            school_year=data["school_year"],
+            class_level=data.get("class_level"),
             class_number=data["class_number"],
             school=school 
         )
@@ -27,10 +21,21 @@ class StudentEnrollmentService:
     def update_enrollment(user, enrollment_id, data):
         
         enrollment = StudentEnrollmentRepository.get_by_id(user, enrollment_id)
+        if enrollment.school != user.counselor.school:
+            raise PermissionError("Not allowed")
+        
+        data.pop("school", None)
+        data.pop("student", None)
+        data.pop("school_year", None)
+
         return StudentEnrollmentRepository.update(enrollment, data)
 
     @staticmethod
     def delete_enrollment(user, enrollment_id):
 
         enrollment = StudentEnrollmentRepository.get_by_id(user, enrollment_id)
+
+        if enrollment.school != user.counselor.school:
+            raise PermissionError("Not allowed")
+        
         StudentEnrollmentRepository.delete(enrollment)
